Add in new game modes for extra chance, guaranteedWildReelsAndWild, BonusBuy and SuperBonusBuy
Added in the basic configuration for these new game modes, the actual quotas, weights, outcome amounts, etc. will still need to be adjusted accordingly to receive desired distribution amounts for how each mode will be able to land.

This is just a base to be able to build off of and adjust accordingly to reach the ultimate desired feel.